Huge optimisation

+++ ArtificialNN/src/Instruments/Functors/Functors.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 #include "Func.h"
-#define DD_Func Func <double, double>
+#define DD_Func Func <float, float>
 #define ID_Func Func <int, double>
 #define DI_Func Func <double, int>
 #define II_Func Func <int, int>
