+++ ArtificialNN/src/Perceptron/Perceptrons.h
@@ -4,7 +4,7 @@
 #include "PLearns.h"
 #include "Weights.h"
 
-#define DD_Perceptron NeyronPerceptron<double, double>
+#define DD_Perceptron NeyronPerceptron<float, float>
 #define ID_Perceptron NeyronPerceptron<int, double>
 #define DI_Perceptron NeyronPerceptron<double, int>
 #define II_Perceptron NeyronPerceptron<int, int>
