+++ ArtificialNN/src/Perceptron/learnPerc/PLearns.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 #include "PerceptronLearning.h"
-#define DD_Leaning PerceptronLearning<double, double>
+#define DD_Leaning PerceptronLearning<float, float>
 #define ID_Leaning PerceptronLearning<int, double>
 #define DI_Leaning PerceptronLearning<double, int>
 #define II_Leaning PerceptronLearning<int, int>
